Reduce all task types to the template
