+++ crates/task/src/lib.rs
@@ -7,7 +7,9 @@ mod vscode_format;
 
 use collections::HashMap;
 use gpui::ModelContext;
-use static_source::RevealStrategy;
+use schemars::gen::SchemaSettings;
+use schemars::JsonSchema;
+use serde::{Deserialize, Serialize};
 use std::any::Any;
 use std::borrow::Cow;
 use std::path::PathBuf;
@@ -77,6 +79,8 @@ impl VariableName {
 impl std::fmt::Display for VariableName {
     fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
         match self {
+            // TODO kb use `ZED_TASK_` prefix for the vars (and `ZED_TASK_CUSTOM_` too?),
+            // extract it into the const and check for substitution issues later
             Self::File => write!(f, "ZED_FILE"),
             Self::WorktreeRoot => write!(f, "ZED_WORKTREE_ROOT"),
             Self::Symbol => write!(f, "ZED_SYMBOL"),
@@ -139,6 +143,132 @@ pub trait Task {
     /// Sets up everything needed to spawn the task in the given directory (`cwd`).
     /// If a task is intended to be spawned in the terminal, it should return the corresponding struct filled with the data necessary.
     fn prepare_exec(&self, cx: TaskContext) -> Option<SpawnInTerminal>;
+}
+
+/// TODO kb proper docs
+#[derive(Clone, Debug, PartialEq)]
+pub struct TaskTemplate {
+    /// TODO kb proper docs
+    pub id: TaskId,
+    /// TODO kb proper docs
+    pub definition: Definition,
+}
+
+impl TaskTemplate {
+    fn oneshot(prompt: String) -> Arc<dyn Task> {
+        Arc::new(Self {
+            id: TaskId(prompt.clone()),
+            definition: Definition {
+                label: prompt.clone(),
+                command: prompt,
+                ..Definition::default()
+            },
+        })
+    }
+}
+
+impl Task for TaskTemplate {
+    fn prepare_exec(&self, cx: TaskContext) -> Option<SpawnInTerminal> {
+        let TaskContext {
+            cwd,
+            task_variables,
+        } = cx;
+        // TODO kb ensure all substitutions are possible to do: no `cwd` has the task prefix, no `env`, `args`, `label`, or `command` have vars with task prefix that are not in `task_variables`. Omit such tasks. + test this
+        let task_variables = task_variables.into_env_variables();
+        let cwd = self
+            .definition
+            .cwd
+            .clone()
+            .and_then(|path| {
+                subst::substitute(&path, &task_variables)
+                    .map(Into::into)
+                    .ok()
+            })
+            .or(cwd);
+        let mut definition_env = self.definition.env.clone();
+        definition_env.extend(task_variables);
+        Some(SpawnInTerminal {
+            id: self.id.clone(),
+            cwd,
+            use_new_terminal: self.definition.use_new_terminal,
+            allow_concurrent_runs: self.definition.allow_concurrent_runs,
+            // TODO kb use expanded label here
+            label: self.definition.label.clone(),
+            command: self.definition.command.clone(),
+            args: self.definition.args.clone(),
+            reveal: self.definition.reveal,
+            env: definition_env,
+        })
+    }
+
+    fn name(&self) -> &str {
+        &self.definition.label
+    }
+
+    fn id(&self) -> &TaskId {
+        &self.id
+    }
+
+    fn cwd(&self) -> Option<&str> {
+        self.definition.cwd.as_deref()
+    }
+}
+
+/// Static task definition from the tasks config file.
+#[derive(Clone, Default, Debug, PartialEq, Eq, Serialize, Deserialize, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub struct Definition {
+    /// Human readable name of the task to display in the UI.
+    pub label: String,
+    /// Executable command to spawn.
+    pub command: String,
+    /// Arguments to the command.
+    #[serde(default)]
+    pub args: Vec<String>,
+    /// Env overrides for the command, will be appended to the terminal's environment from the settings.
+    #[serde(default)]
+    pub env: HashMap<String, String>,
+    /// Current working directory to spawn the command into, defaults to current project root.
+    #[serde(default)]
+    pub cwd: Option<String>,
+    /// Whether to use a new terminal tab or reuse the existing one to spawn the process.
+    #[serde(default)]
+    pub use_new_terminal: bool,
+    /// Whether to allow multiple instances of the same task to be run, or rather wait for the existing ones to finish.
+    #[serde(default)]
+    pub allow_concurrent_runs: bool,
+    /// What to do with the terminal pane and tab, after the command was started:
+    /// * `always` — always show the terminal pane, add and focus the corresponding task's tab in it (default)
+    /// * `never` — avoid changing current terminal pane focus, but still add/reuse the task's tab there
+    #[serde(default)]
+    pub reveal: RevealStrategy,
+}
+
+/// What to do with the terminal pane and tab, after the command was started.
+#[derive(Default, Clone, Copy, Debug, PartialEq, Eq, Serialize, Deserialize, JsonSchema)]
+#[serde(rename_all = "snake_case")]
+pub enum RevealStrategy {
+    /// Always show the terminal pane, add and focus the corresponding task's tab in it.
+    #[default]
+    Always,
+    /// Do not change terminal pane focus, but still add/reuse the task's tab there.
+    Never,
+}
+
+/// A group of Tasks defined in a JSON file.
+#[derive(Clone, Debug, Default, PartialEq, Eq, Serialize, Deserialize, JsonSchema)]
+pub struct TaskDefinitions(pub Vec<Definition>);
+
+impl TaskDefinitions {
+    /// Generates JSON schema of Tasks JSON definition format.
+    pub fn generate_json_schema() -> serde_json_lenient::Value {
+        let schema = SchemaSettings::draft07()
+            .with(|settings| settings.option_add_null_type = false)
+            .into_generator()
+            .into_root_schema_for::<Self>();
+
+        serde_json_lenient::to_value(schema).unwrap()
+    }
 }
 
 /// [`Source`] produces tasks that can be scheduled.
